% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggamma.R
\name{expansionControl}
\alias{expansionControl}
\title{Control Parameters for dggamma and nllggamma}
\usage{
expansionControl(...)
}
\arguments{
\item{...}{Named parameters to override defaults.}
}
\value{
A list with control parameters.
}
\description{
Creates a list of control parameters for \code{dggamma} and \code{nllggamma}.
}
\examples{
expansionControl(epsilon = c(1e-2, 1/3), kappa = 10L)
}
