% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggamma.R
\name{nllggamma}
\alias{nllggamma}
\title{Negative Log-Likelihood for Generalized Gamma}
\usage{
nllggamma(p, x, links, w = 1, control = expansionControl())
}
\arguments{
\item{p}{A numeric vector of parameters.}

\item{x}{Numeric vector of data.}

\item{links}{Character vector (either "identity", "log", or "logit") specifying the three link functions for parameters.}

\item{control}{A list of control parameters, typically created using \code{expansionControl()}.}
}
\value{
The negative log-likelihood, with attributes for the gradient and Hessian matrix.
}
\description{
Computes the negative log-likelihood (w/ gradient and Hessian) for the generalized gamma distribution.
}
\details{
This function is meant to be used with nlm to fit an extended generalized distribution.

The parameters \code{p} are those obtained by applying the link functions to mu, sigma, and nu.
In other words, using \code{links = rep("identity", 3)} implies that \code{p = c(mu, sigma, nu)}.
Default is \code{links = c("log", "log", "identity")}.

The control parameters are \code{epsilon} and \code{kappa}. \code{epsilon} is a length two numeric vector
indicating the intervals \code{c(-epsilon[2],epsilon[1])} and \code{c(epsilon[1],epsilon[2])} where the
transition is made from the standard generalized gamma negative log-lik to its approximation. \code{kappa}
determines the maximal power of \code{nu} to consider in the Taylor and Stirling expansions formulas.
}
\examples{
nllggamma(p = c(1, 0, 0), x = c(1, 2, 3), links = c("identity", "log", "logit"))
}
