% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggamma.R
\name{qggamma}
\alias{qggamma}
\title{Generalized Gamma Quantile Function}
\usage{
qggamma(
  p,
  mu = 1,
  sigma = 0.5,
  nu = 1,
  lower.tail = TRUE,
  log.p = FALSE,
  epsilon = 1e-04
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{mu}{Positive numeric, location parameter.}

\item{sigma}{Positive numeric, scale parameter.}

\item{nu}{Numeric, shape parameter.}

\item{lower.tail}{Logical; defined as usual.}

\item{log.p}{logical; if \code{TRUE}, probabilities \code{p} are returned as \code{log(p)}.}

\item{epsilon}{Positive number defining when to start using linear interpolation (from \code{abs(nu)=epsilon} to \code{nu=0}).}
}
\value{
A numeric vector of quantiles.
}
\description{
Computes the quantiles of the generalized gamma distribution using a simple linear interpolation around nu=0.
}
